Redefining fstype as the internal fs type, original fstype is still available using statfstype.
Thanks @martinvw for the inspiration, but his approach removed the fstype as determined by the statistics entirely after get_fs_type was ran, I think this is a bit clearer. It also makes sure that the string representation of Volume does not change and keeps fill_stats optional.

This is technically a backwards incompatible change, but it should not break anything. Just don't use get_fs_type anymore, but access Volume.fstype instead.

This commit also changes the way determine_fs_type works: as there are currently two sources for fs type information, we now loop over the available sources and try to find a match. This removes the need to list 'generic' fs types (the old list is still included to fail fast).