Increase the allowed file extension length for segmented images.
This commit increases the allowed file extension length
for segmented images from 3 to 4 characters.

It should be noted that the file extensions should be of a
consistent length for a set of segmented file otherwise
the sorting internally in various mounting methods (eg xmount)
will not work correctly and the resultant image will be corrupt.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>