Allow longer numerical file extensions for segmented images
This commit allows longer numerical file extensions. Instead
of limiting numical file extensions to 2 or 3 characters,
we can now handle numerical file extensions of >=2 characters.

As with the previous commit, file extensions should be of a
consistent length for a set of segmented files otherwise the
sorting internally in various mounting methods (eg xmount)
will not work correctly and the resultant image will be corrupt.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>