Change how fstype dependencies are resolved, fixes #36
The detection method of filesystems was flawed. This is fixed by checking the existence of the fstype in /proc/filesystems, as fs- module or as mount.type script.

This commit also optimizes the tests in full_image_test.py to make use of this new detection method and use pytest paradigms.